fix(ARC-1c): fix lint errors and extract types to stay under 500 lines
